refactor(literature): split LiteratureSearch.tsx into page + 8 modules
2385 lines in one component is not reviewable, which is the point mrj4001
made on the PR. The page is now 968 lines and keeps every hook it had.

Extracted, in order of how safe each was:

  LiteratureSearch.types.ts      Seed, DbResult, DbFailure, Expert, isFailure
  LiteratureSearch.constants.ts  MODES, DATABASES, SORTS, STAGES, ROWS_RETRY_MS
  LiteratureSearch.parts.tsx     RowCount, LineInput, Prose, ProgressPanel,
                                 ExpertsPanel, FailedDatabases, PageHead, SummaryBar
  LiteratureSearch.downloads.ts  makeDownloads(ctx) — the whole DOWNLOADS block
  SearchForm.tsx                 the form as 7 field-level components
  StrategyResults.tsx            Mode 1 panels + seed validation
  CandidatesView.tsx             query card, narrowing gate, record rows
  SynthesisView.tsx              Mode 2/3 synthesis

Behaviour-neutral. No state moved, no hook order changed, no dependency
array touched, no CSS class renamed. Every extracted component is declared
at module level, so nothing remounts. The default export and its path are
unchanged, so no import site anywhere moved.

What did NOT come out, and why: all page state, the four POST handlers, the
debounced re-count effect and every strategy mutator each call a setter, so
extracting them would mean lifting or splitting state. That block is ~640 of
the remaining 968 lines.

Also removes four dead locals (liveStrategy, retrieved, allFound, numbered)
that were shadowed by identically-named consts inside the per-database map.
Nothing ever read the outer ones; leaving them would have made them look live
once their shadowing twins moved to another file.
